@@ -1,6 +1,5 @@
 <?php
 
-use App\Http\Controllers\AuthController;
 use App\Http\Controllers\HomeController;
 use App\Http\Controllers\Influenceur\DashboardController;
 use App\Http\Controllers\Marque\DashboardController as MarqueDashboardController;
@@ -17,52 +16,20 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::name('home.')->group(function () {
-    Route::get('/', [HomeController::class, 'index'])->name('index');
-   
-    Route::get('/register', [HomeController::class, 'register'])->name('register');
-    Route::post('/register', [HomeController::class, 'store'])->name('register');
+Route::get('/', [HomeController::class, 'index'])->name('home');
+Route::get('profile', fn () => view('profile'))->name('profile');
+
+Route::get('/partnership', fn () => view('partnership'))->name('partnership');
+Route::get('/inbox', fn () => view('inbox'))->name('inbox');
+Route::get('/earnings', fn () => view('earnings'))->name('earnings');
+Route::get('/search', fn () => view('results'))->name('search.results');
+
+Route::middleware([
+    'auth:sanctum',
+    config('jetstream.auth_session'),
+    'verified',
+])->group(function () {
+    Route::get('/dashboard', function () {
+        return view('dashboard');
+    })->name('dashboard');
 });
-
-Route::name('auth.')->group(function () {
-    Route::get('/login', [AuthController::class, 'loginForm'])->name('login');
-    Route::post('/login', [AuthController::class, 'login'])->name('login');
-    Route::delete('/logout', [AuthController::class, 'logout'])->name('logout');
-});
-
-
-Route::prefix('influenceur')
-    ->middleware('auth')
-    ->name('influenceur.')
-    ->controller(DashboardController::class)
-    ->group( function () {
-   
-    Route::get('/', 'index')->name('index');
-    Route::get('/Partenariats', 'partenariat')->name('partenariat');
-    Route::get('/Boite_de_reception', 'message')->name('message');
-    Route::get('/Gains', 'gain')->name('gain');
-    Route::get('/Profil', 'profil')->name('profil');
-    
-});
-
-
-Route::prefix('marque')
-    ->middleware('auth')
-    ->name('marque.')
-    ->controller(MarqueDashboardController::class)
-    ->group( function () {
-   
-    Route::get('/', 'index')->name('index');
-    Route::get('/Partenariats', 'partenariat')->name('partenariat');
-    Route::get('/Boite_de_reception', 'message')->name('message');
-    Route::get('/Profil', 'profil')->name('profil');
-
-    Route::get('/Resultat_de_recherche', 'search')->name('search');
-    
-});
-
-
-
-
-
-
