@@ -22,7 +22,7 @@ class SearchRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'key' => 'required'
+            'query' => 'required|string'
         ];
     }
 }
